Add broker support for passing app_name
In Ceph >= Luminous, application name needs to be set
on a per-pool level to avoid health warnings. This
change adds support for sending the application name
over the broker channel from consuming charms.

When a name is not sent from the other side of the
relation, the application name will be set to "unknown"
in Luminous and greater

Change-Id: I1109251b08da20adaf3d677c38fc1aacfba29439
Closes-Bug: #1753640
Depends-On: I99ae47b6802f50ea019751ffa328f11567cca567