Remove explicit fsid charm config option
The charm has had the capability to automatically handle fsid
in deployments for several cycles.

Explicitly setting the fsid can lead to unpredictable behavior
when data from previous deployments remains on block devices,
and when the fsid is explicitly static across those separate
deployments.  ex. Ceph would see the fsid match and attempt to
use the block device from a previous deployment as-is and
unsuccessfully.

This primarily affects legacy development and CI environments,
which may be in place from a time before the charm autoatically
handled fsids.  The recommendation is for both test and production
deployments to defer fsid generation to the charm.

Change-Id: I0b87576810faa08a81dc2d559ef925ea02f58db0
Partial-bug: #1698154