Guard cluster operations until sufficient OSD's booted
Ensure that broker requests are not processed and that client
access keys are not issued until the expected number of OSD's
have been bootstrapped into the cluster.

This depends on presentation of the number of bootstrapped
OSD's from the ceph-osd charm (see Depends-On).

For upgraders, keys will have already been issued so there
should be no impact on existing access to the Ceph cluster;
The ceph-osd units will present the required relation data
post upgrade at which point the charm will mark the cluster
as ready for service and continue to process and pending
requests.

Change-Id: Id67e13c176fc8fd4953ba7c2cf7e33252810940c
Closes-Bug: 1794878
(cherry picked from commit ed99bd2b5708d0494a9139e6805a03940b41eefb)