Disable object skew warnings
Ceph will issue a HEALTH_WARN in the event that one pool has a
large number of objects compared to other pools in the cluster:

 "Issue a HEALTH_WARN in cluster log if the average object
  number of a certain pool is greater than mon pg warn max
  object skew times the average object number of the whole
  pool."

For OpenStack deployments, Gnocchi and RADOS gateway can generate
a large number of small objects compared to Cinder, Glance and
Nova usage, causing the cluster to go into HEALTH_WARN status.

Disable this check until the skew evaluation also includes the
size of the objects as well as the number.

Change-Id: I83211dbdec4dea8dca5b27a66e26a4431d2a7b77
Closes-Bug: 1804846