Update cloud-archive.list when upgrading from Pike to Queens
ceph-mon charm only upgrades when the ceph version changes, for the
case of upgrading from Pike to Queens the charm is skipping any
upgrades, because the Cloud Archive has Luminous for those 2 releases.

This patch checks if the requested ceph version is luminous and if the
'source' changed from pike to queens to then upgrade
/etc/apt/sources.list.d/cloud-archive.list via add_source()

As part of this backport, I manually un-picked the stein dict
entry from lib/ceph/utils.py, as it is not currently applicable to
the stable charm.

Change-Id: I05b7d722e45d3a02a97866903a67bd9b16d4f552
Closes-Bug: 1778823
(cherry picked from commit 5ed4956f2bc594d42abf6fd2fa64dcddc1b86da5)