Ensure relation id is passed to remote_service_name
Its possible for the client_relation handler to be called from
outside of the original hook context; as a result use of
remote_service_name is not safe as this will evaluate in the
current hook context, rather than the relation_id/unit context
provided in the function parameters.

Passing the relation_id to the helper ensures that the
remote units for the intended relation are used to generate
the remote service name.

Change-Id: I2305a111c0a618156208e4f96acda5e63736051c
Closes-Bug: 1819892