Creates aditional nrpe checks which parse warning/error messages
When Ceph is in a warning state for reason1 and in the meantime
new reason2 appears, operator is not alerted and also cannot mute
alarms selectively (as described in bug #1735579)
This patch alllows to specify a dictionary of 'name':'regex' where
'name' will become ceph-$name check in nrpe and $regex will be
searched for in warning/error messages. It is specified via
charm nagios_additional_checks parameter.
There is also nagios_additional_checks_critical parameter which
specifies if those checks are reported as warning or error.

Change-Id: I73a7c15db88793bb78841d8395535c97ca2af872
Partial-Bug: 1735579