Comprehensive review of ceph-mon README
Review of README.

Actions section was added.

The removed upgrade information will be moved
to the CDG upgrade-openstack appendix.

The upgrade paths information I don't believe
is necessary. The UCA method is already generically
laid out in the CDG upgrade appendix. The only
benefit is for the user to know what version of
Ceph they will end up with, but this is hard
to keep up to date in our documentation.

The upgrade edge case section will be moved to the
upgrade issues section of the upgrade appendix.

Change-Id: Ie4d8d69dc400cc79aa966bd72d7c601882a50bfc