Set sane min pg size with autoscaler
Resync charmhelpers to pickup behavioural changes in
autoscaler configuration for pools.

This ensures that pools that have a calculated pg size
less than the default of 32 get a min size set to match
the calculated pg size avoiding instance scale-up to
32 pgs.

This avoids overloading of the max pgs per osd limit
in smaller test clusters.

Change-Id: Ic34d029dae9c67dbba3e2e502d7c9ac4576fcfa5
Closes-Bug: 1872748