Add erasure-coded pool bundle
This change adds a bundle that configures erasure coded pools
for glance, cinder-ceph and nova. To verify that the pools are
using EC it also adds a test which verifies the pool type.

To verify that the nova pool is configured correctly in all
scenarious all the bundles have been updated to set
libvirt-image-backend to rbd which triggers the nova pool
creation.

During testing the Prometheus tests were failing which is
down to a known bug *1 in the prometheus charm. The
bundles have been updated to pin the prometheus2 charm
to an earlier version to get round the bug.

*1 https://bugs.launchpad.net/charm-prometheus2/+bug/1891942


Change-Id: I7d29aec3ba9924d5677be25137907b1465401f9c