Only check for expected-osd-count until it has been met
Checking for enough OSDs to be presented by OSD units is
necessary during deploy time to ensure that clients can
correctly connect and perform their operations; however,
checking is useless post-deploy and can be harmful during
replacement operations.

This change introduces a restriction that this check should
only be done on the leader, as well as ensuring that it
short-circuits the check after the check passes.

Closes-Bug: #1938970
Change-Id: Ie285bbc34692964acb35315f866fe617b0ef1305